@@ -12,29 +12,32 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.7",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.7",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^7.0.1",
+    "@angular/cdk": "^7.0.1",
+    "@angular/common": "^7.0.1",
+    "@angular/compiler": "^7.0.1",
+    "@angular/core": "^7.0.1",
+    "@angular/forms": "^7.0.1",
+    "@angular/http": "^7.0.1",
+    "@angular/material": "^7.0.1",
+    "@angular/platform-browser": "^7.0.1",
+    "@angular/platform-browser-dynamic": "^7.0.1",
+    "@angular/router": "^7.0.1",
     "@ngrx/core": "^1.2.0",
     "@ngrx/effects": "^6.1.0",
     "@ngrx/store": "^6.1.0",
     "core-js": "^2.5.4",
-    "rxjs": "^6.2.2",
+    "logrocket": "^0.6.17",
+    "logrocket-ngrx": "^0.1.1",
+    "moment": "^2.22.2",
+    "rxjs": "^6.3.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.8.3",
-    "@angular/cli": "~6.2.1",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "~7.0.3",
+    "@angular/compiler-cli": "^7.0.1",
+    "@angular/language-service": "^7.0.1",
     "@ngrx/store-devtools": "^6.1.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
@@ -50,6 +53,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "~3.1.3"
   }
 }
