@@ -16,40 +16,111 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>
  ******************************************************************************/
 
-import {Injectable} from '@angular/core';
-import {Actions, Effect, ofType} from '@ngrx/effects';
-import {map, mergeMap, switchMap} from 'rxjs/operators';
-import {HttpClient} from '@angular/common/http';
-import {from} from 'rxjs';
-import {Router} from '@angular/router';
-import {SET_TOKEN, TRY_SIGNIN, TrySignin} from './auth.actions';
-import {LOGIN_URL} from '../../../statics/api_urls.statics';
+import { Injectable } from "@angular/core";
+import { Actions, Effect, ofType } from "@ngrx/effects";
+import { catchError, map, mergeMap, switchMap } from "rxjs/operators";
+import { HttpClient } from "@angular/common/http";
+import { from } from "rxjs";
+import { Router } from "@angular/router";
+import {
+    TRY_RELOAD_STATIC_INFORMATION,
+    SET_TOKEN,
+    TRY_LOGIN,
+    TryLogin
+} from "./auth.actions";
+import LogRocket from "logrocket";
+import { LOGIN_URL } from "../../../statics/api_urls.statics";
+import { SET_USER } from "../api.actions";
+import { AuthGuardService } from "../../services/auth-guard.service";
+import { FullUser } from "../../models/user.model";
+import {RecordsSandboxService} from '../../../recordmanagement/services/records-sandbox.service';
+import {ApiSandboxService} from '../../services/api-sandbox.service';
 
 @Injectable()
 export class AuthEffects {
-    constructor(private actions: Actions, private http: HttpClient, private router: Router){}
+    constructor(
+        private actions: Actions,
+        private http: HttpClient,
+        private router: Router,
+        private guard: AuthGuardService,
+        private recordSB: RecordsSandboxService,
+        private apiSB: ApiSandboxService
+    ) {}
 
     @Effect()
-    authSignin = this.actions.pipe(
-        ofType(TRY_SIGNIN),
-        map((action: TrySignin) => {
+    tryLogin = this.actions.pipe(
+        ofType(TRY_LOGIN),
+        map((action: TryLogin) => {
             return action.payload;
         }),
-        switchMap((authData: {username: string, password: string}) => {
-            return from(this.http.post(LOGIN_URL, authData));
+        switchMap((authData: { username: string; password: string }) => {
+            return from(
+                this.http.post(LOGIN_URL, authData).pipe(
+                    catchError(error => {
+                        console.log('error', error)
+                        this.apiSB.showErrorSnackBar("Login not successfull");
+                        return [];
+                    }),
+                    mergeMap(
+                        (response: {
+                            token: string;
+                            user: any;
+                            consultants: any;
+                            countries: any;
+                        }) => {
+                            localStorage.setItem("token", response.token);
+
+                            if (this.guard.lastVisitedUrl)
+                                this.router.navigate([
+                                    this.guard.getLastVisitedUrl()
+                                ]);
+                            else this.router.navigate([""]);
+
+                            return [
+                                {
+                                    type: SET_TOKEN,
+                                    payload: response.token
+                                },
+                                ...AuthEffects.getStaticInformation(response)
+                            ];
+                        }
+                    )
+                )
+            );
         }),
-        mergeMap((response: {token: string}) => {
-            console.log(response);
-            localStorage.setItem("token", response.token);
-            this.router.navigate([""]);
-            return [
-                {
-                    type: SET_TOKEN,
-                    payload: response.token
-                }
-            ];
+    );
+
+    @Effect()
+    setToken = this.actions.pipe(
+        ofType(SET_TOKEN),
+        switchMap(() => {
+            this.recordSB.startLoadingRecordStatics();
+            return [];
         })
     );
 
+    @Effect()
+    reload = this.actions.pipe(
+        ofType(TRY_RELOAD_STATIC_INFORMATION),
+        switchMap(() => {
+            return from(this.http.get(LOGIN_URL));
+        }),
+        mergeMap((response: any) => {
+            return [...AuthEffects.getStaticInformation(response)];
+        })
+    );
 
+    static getStaticInformation(response: { user: any; }) {
+        // for logging
+        LogRocket.identify(response.user.id);
+        // keep this console.log
+        console.log("identified: ", response.user.id);
+
+        return  [
+            {
+                type: SET_USER,
+                payload: FullUser.getFullUserFromJson(response.user)
+            }
+        ];
+    }
 }
