@@ -24,7 +24,7 @@ class UpdateOwnProfile(permissions.BasePermission):
     def has_object_permission(self, request, view, obj):
         """check user is trying to edit their own profile"""
 
-        if request.method in permissions.SAFE_METHODS:
+        if request.method in permissions.SAFE_METHODS or request.user.is_superuser:
             return True
 
         return obj.id == request.user.id or request.user.is_staff
@@ -37,7 +37,7 @@ class EditRecord(permissions.BasePermission):
 
 class TestIt(permissions.BasePermission):
     def has_permission(self, request, view):
-        a = request.method
+        # TODO: remove this class
         return True
 
 
