@@ -24,5 +24,9 @@ router.register('origin_countries', OriginCountriesViewSet)
 router.register('record_tags', RecordTagViewSet)
 
 urlpatterns = [
-  url(r'', include(router.urls)),
+    url(r'', include(router.urls)),
+    url(r'statics', StaticViewSet.as_view()),
+    url(r'clients_by_birthday', GetClientsFromBirthday.as_view()),
+    url(r'record/(?P<id>.+)/$', RecordViewSet.as_view()),
+    url(r'record', RecordViewSet.as_view()),
 ]
