@@ -8,11 +8,13 @@ https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
 """
 
 import os
-
+import sys
 from django.core.wsgi import get_wsgi_application
 from whitenoise.django import DjangoWhiteNoise
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'backend.settings')
+os.environ["DJANGO_SETTINGS_MODULE"] = 'backend.settings'
 
 application = get_wsgi_application()
 application = DjangoWhiteNoise(application)
+
