@@ -16,35 +16,81 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>
  ******************************************************************************/
 
-export class FullUser {
-    public id: string;
-    public email: string;
-    public name: string;
-    public birthday: Date;
-    public phone_number: string;
-    public street: string;
-    public city: string;
-    public postal_code: string;
+import {ApiSandboxService} from '../services/api-sandbox.service';
+import {Filterable} from '../../shared/models/filterable.model';
+
+export class RestrictedUser implements Filterable{
+    constructor(public id: string, public name: string) {
+        this.id = id;
+        this.name = name;
+    }
+
+    static getRestrictedUsersFromJsonArray(jsonArray) {
+        const restrictedUsers: Array<RestrictedUser> = [];
+        Object.values(jsonArray).map(restrictedJsonUser => {
+            restrictedUsers.push(RestrictedUser.getRestrictedUserFromJson(restrictedJsonUser));
+        });
+        return restrictedUsers;
+    }
+
+    static getRestrictedUserFromJson(json){
+        return new RestrictedUser(json.id, json.name);
+    }
+
+    getFilterableProperty(){
+        return this.name;
+    }
+}
+
+export class FullUser extends RestrictedUser {
 
     constructor(
-        id: string = '',
-        email: string = '',
-        name: string = '',
-        birthday: Date = new Date(),
-        phone_number: string = '',
-        street: string = '',
-        city: string = '',
-        postal_code: string = ''
+        id: string = "",
+        public email: string = "",
+        name: string = "",
+        public birthday: Date = new Date(),
+        public phone_number: string = "",
+        public street: string = "",
+        public city: string = "",
+        public postal_code: string = ""
     ) {
-        this.id = id;
+        super(id, name);
         this.email = email;
-        this.name = name;
         this.birthday = birthday;
         this.phone_number = phone_number;
         this.street = street;
         this.city = city;
         this.postal_code = postal_code;
     }
-}
 
-export class RestrictedUser {}
+    static getFullUserFromJson(json) {
+        return new FullUser(
+            json.id,
+            json.email,
+            json.name,
+            new Date(json.birthday),
+            json.phone_number,
+            json.street,
+            json.city,
+            json.postal_code
+        );
+    }
+
+    /**
+     * compares this instance with another and returns updateObject with every field which will be changed
+     * used to generated the update object for patching user in backend
+     * @param updates potential updates for object
+     */
+    getUpdates(updates: FullUser) {
+        const changes = {};
+        if (this.birthday !== updates.birthday)
+            changes["birthday"] = ApiSandboxService.transformDate(updates.birthday);
+        if (this.phone_number !== updates.phone_number)
+            changes["phone_number"] = updates.phone_number;
+        if (this.street !== updates.street) changes["street"] = updates.street;
+        if (this.city !== updates.city) changes["city"] = updates.city;
+        if (this.postal_code !== updates.postal_code)
+            changes["postal_code"] = updates.postal_code;
+        return changes;
+    }
+}
