@@ -19,42 +19,73 @@
 export class RestrictedRecord {
     constructor(
         public id: number,
+        public token: string,
         public last_contact_date: Date,
         public state: string,
         public tags: [number, string],
-        public working_on_record: [number]
-    ){
+        public working_on_record: [number, string]
+    ) {
         this.id = id;
+        this.token = token;
         this.last_contact_date = last_contact_date;
         this.state = state;
         this.tags = tags;
         this.working_on_record = working_on_record;
     }
+
+    static getRestrictedRecordFromJson(json){
+        return new RestrictedRecord(
+            json.id,
+            json.record_token,
+            new Date(json.last_contact_date),
+            json.state,
+            json.tagged,
+            json.working_on_record
+        );
+    }
 }
 
-export class FullRecord extends RestrictedRecord{
+export class FullRecord extends RestrictedRecord {
     constructor(
         id: number,
+        token: string,
         last_contact_date: Date,
         state: string,
         tags: [number, string],
-        working_on_record: [number],
+        working_on_record: [number, string],
         public created_on: Date,
         public last_edited: Date,
         public first_contact_date: Date,
-        public record_token: string,
+        //public record_token: string,
         public note: string,
         public from_rlc: number,
-        public client: number,
+        public client: number
     ) {
-        super(id, last_edited, state, tags, working_on_record);
+        super(id, token, last_edited, state, tags, working_on_record);
         this.created_on = created_on;
         this.last_edited = last_edited;
         this.first_contact_date = first_contact_date;
-        this.record_token = record_token;
+        //this.record_token = record_token;
         this.note = note;
         this.from_rlc = from_rlc;
         this.client = client;
     }
-}
 
+    static getFullRecordFromJson(json){
+        return new FullRecord(
+            json.id,
+            json.record_token,
+            new Date(json.last_contact_date),
+            json.state,
+            json.tagged,
+            json.working_on_record,
+            new Date(json.created_on),
+            new Date(json.last_edited),
+            new Date(json.first_contact_date),
+            //json.record_token,
+            json.note,
+            json.from_rlc,
+            json.client
+        );
+    }
+}
