@@ -16,5 +16,23 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>
  ******************************************************************************/
 
-export const LOGIN_URL = "/api/login/";
-export const RECORDS_URL = 'api/records/records/';
+export const LOGIN_URL = "api/login/";
+export const RECORDS_URL = "api/records/records/";
+const RECORD_SPECIAL_URL = "api/records/record/";
+export const CLIENTS_BY_BIRTHDAY_URL = "api/records/clients_by_birthday/";
+export const RECORDS_STATICS_URL = "api/records/statics/";
+export const PROFILES_URL = "api/profiles/";
+export const CREATE_PROFILE_URL = "api/create_profile/";
+export const CREATE_RECORD_URL = "api/records/create_record/";
+export const RLCS_URL = "api/get_rlcs/";
+
+export function GetSpecialProfileURL(id: string | number){
+    return `${PROFILES_URL}${id}/`;
+}
+export function GetRecordsSearchURL(toSearch: string){
+    return `${RECORDS_URL}?search=${toSearch}`;
+}
+
+export function GetSpecialRecordURL(id: string | number){
+    return `${RECORD_SPECIAL_URL}${id}/`;
+}
