@@ -60,7 +60,7 @@
 
 /** Evergreen browsers require these. **/
 // Used for reflect-metadata in JIT. If you use AOT (and only Angular decorators), you can remove.
-import "core-js/es7/reflect";
+
 
 /**
  * Web Animations `@angular/platform-browser/animations`
